+++ extension/src/components/vanillaTheme/constants.ts
@@ -1,0 +1,29 @@
+/**
+ * ========================================================================================
+ * Humbanew Project ©️ 2023-2025
+ * All rights reserved.
+ * Licensed under the MIT License, adapted.
+ * ========================================================================================
+ */
+
+import { VscodeColorTheme } from "./defines";
+
+export const THEME_ADVANCED_CIRCUITS = (): VscodeColorTheme["colors"] => { return {} }
+,THEME_BLUEPRINT_PAPER = (): VscodeColorTheme["colors"] => { return {} }
+,THEME_HUMBA01_DESIGN_STYLE = (): VscodeColorTheme["colors"] => { return {} }
+,THEME_WINTER_DAY = (): VscodeColorTheme["colors"] => { return {} }
+,THEME_CAPPUCCINO = (): VscodeColorTheme["colors"] => { return {} }
+,THEME_RED_VELVET = (): VscodeColorTheme["colors"] => { return {} }
+,THEME_SUNSHINE = (): VscodeColorTheme["colors"] => { return {} }
+,THEME_RI_VISUAL_STUDIO = (): VscodeColorTheme["colors"] => { return {} }
+,THEME_RI_VISUAL_STUDIO_BLUE = (): VscodeColorTheme["colors"] => { return {} }
+,HIGHLIGHT_VISUAL_STUDIO_CODE = (): VscodeColorTheme["tokenColors"] => { return [] }
+,HIGHLIGHT_MONOKAI = (): VscodeColorTheme["tokenColors"] => { return [] }
+,HIGHLIGHT_MONOKAI_DIMMED = (): VscodeColorTheme["tokenColors"] => { return [] }
+,HIGHLIGHT_BLULOCO = (): VscodeColorTheme["tokenColors"] => { return [] }
+,HIGHLIGHT_BLULOCO_ITALIC = (): VscodeColorTheme["tokenColors"] => { return [] }
+,HIGHLIGHT_DRACULA = (): VscodeColorTheme["tokenColors"] => { return [] }
+,HIGHLIGHT_DRACULA_SOFT = (): VscodeColorTheme["tokenColors"] => { return [] }
+,HIGHLIGHT_ONE_DARK = (): VscodeColorTheme["tokenColors"] => { return [] }
+,HIGHLIGHT_ONE_DARK_VIVID = (): VscodeColorTheme["tokenColors"] => { return [] }
+,HIGHLIGHT_ORIGINAL = (): VscodeColorTheme["tokenColors"] => { return [] };